@@ -1,14 +1,20 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { PageNotFoundComponent }  from './page-not-found.component';
-import { WeatherSearchComponent } from './weather/weather-search.component';
+import { WeatherSearchComponent } from './weather/weather-search/weather-search.component';
 import { WeatherTodayComponent } from './weather-today/weather-today.component';
+import { SigninComponent } from './auth/signin/signin.component';
+import { SignupComponent } from './auth/signup/signup.component';
+import { AuthGuard } from './auth/auth-guard.service';
+import { SignoutComponent } from './auth/signout/signout.component';
 
 const routes: Routes = [
-  //{ path: '', component: WeatherSearchComponent},
   { path: 'today', component: WeatherTodayComponent},
-  { path: 'forecast', component: WeatherSearchComponent },
-  { path: '', redirectTo: '/forecast', pathMatch: 'full' },
+  { path: 'forecast', component: WeatherSearchComponent, canActivate: [AuthGuard]},
+  { path: 'signin', component: SigninComponent },
+  { path: 'signup', component: SignupComponent },
+  { path: 'signout', component: SignoutComponent },
+  { path: '', redirectTo: '/signin', pathMatch: 'full' },
   { path: '**', component: PageNotFoundComponent }
 ];
 
@@ -17,6 +23,3 @@ const routes: Routes = [
   exports: [ RouterModule ]
 })
 export class AppRoutingModule { }
-
-
-
