@@ -1,11 +1,11 @@
-import { InjectionToken } from "@angular/core";
+import { InjectionToken } from '@angular/core';
 
-export let APP_CONFIG = new InjectionToken("app.config");
+export let APP_CONFIG = new InjectionToken('app.config');
 
 export interface IAppConfig {
     apiEndpoint: string;
 }
 
 export const AppConfig: IAppConfig = {
-    apiEndpoint: "http://localhost:15422/api/"
+    apiEndpoint: 'http://localhost:15422/api/'
 };
