@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-weather-today',
   templateUrl: './weather-today.component.html',
-  styleUrls: ['./weather-today.component.css']
+  styleUrls: ['./weather-today.component.scss']
 })
 export class WeatherTodayComponent implements OnInit {
 
