@@ -1,10 +1,10 @@
-export class Weather{
-   constructor(
-               public city: string, 
-               public description: string,
-               public temperature: number){
-                     this.city = city;
-                     this.description = description;
-                     this.temperature = temperature;
-               } 
- }
+export class Weather {
+    constructor(
+        public city: string,
+        public description: string,
+        public temperature: number) {
+        this.city = city;
+        this.description = description;
+        this.temperature = temperature;
+    }
+}
